server/WhiteMarket/requirements.txt
Django===2.0.3
django-cors-headers===2.2.0
djangorestframework===3.7.7
djoser===1.1.5
djangorestframework-jwt===1.11.0
django-filter===1.1.0
psycopg2===2.7.4
psycopg2-binary===2.7.4
Pillow===5.1.0